+++ wsi_service/slide_manager.py
@@ -20,6 +20,11 @@ class SlideManager:
         self.lock = asyncio.Lock()
         self.storage_locks = {}
         self.event_loop = asyncio.get_event_loop()
+        self.local_mapper = None
+
+    def with_local_mapper(self, local_mapper):
+        self.local_mapper = local_mapper
+        return self
 
     async def get_slide(self, slide_id, plugin=None):
         cache_id = slide_id + f" ({plugin})" if plugin else slide_id
@@ -87,18 +92,27 @@ class SlideManager:
         logger.debug("Set expiration timer for storage address (%s): %s", cache_id, self.timeout)
 
     async def _get_slide_storage_addresses(self, slide_id):
-        try:
-            async with aiohttp.ClientSession() as session:
-                async with session.get(self.mapper_address.format(slide_id=slide_id)) as r:
-                    if r.status == 404:
-                        raise HTTPException(
-                            status_code=404, detail=f"Could not find a storage address for slide id {slide_id}."
-                        )
-                    slide = await r.json()
-        except aiohttp.ClientConnectorError:
-            raise HTTPException(
-                status_code=503, detail="WSI Service is unable to connect to the Storage Mapper Service."
-            )
+        slide = None
+        if self.local_mapper:
+            slide = self.local_mapper.get_slide(slide_id)
+            if not slide:
+                raise HTTPException(
+                    status_code=404, detail=f"Could not find a storage address for slide id {slide_id}."
+                )
+            slide = slide.slide_storage.model_dump()
+        else:
+            try:
+                async with aiohttp.ClientSession() as session:
+                    async with session.get(self.mapper_address.format(slide_id=slide_id)) as r:
+                        if r.status == 404:
+                            raise HTTPException(
+                                status_code=404, detail=f"Could not find a storage address for slide id {slide_id}."
+                            )
+                        slide = await r.json()
+            except aiohttp.ClientConnectorError:
+                raise HTTPException(
+                    status_code=503, detail="WSI Service is unable to connect to the Storage Mapper Service."
+                )
         return slide["storage_addresses"]
 
     async def _get_slide_main_storage_address(self, slide_id):