+++ wsi_service_base_plugins/openslide/wsi_service_plugin_openslide/__init__.py
@@ -10,9 +10,10 @@ def is_supported(filepath):
         suffix = pathlib.Path(filepath).suffix
         if filename.endswith("ome.tif") or filename.endswith("ome.tiff"):
             return False
-        return suffix in [".bif", ".mrxs", ".ndpi", ".scn", ".svs", ".tiff", ".tif"]
-    else:
-        return any(list(pathlib.Path(filepath).glob("*.vsf")))
+        return suffix in [".bif", ".mrxs", ".ndpi", ".scn", ".svs", ".tiff", ".tif", ".czi"]
+    # VSF is not supported
+    # else:
+    #     return any(list(pathlib.Path(filepath).glob("*.vsf")))
 
 
 async def open(filepath):